Add files via upload
TASK 1
A school keeps records of the weights of each pupil. The weight, in kilograms, of each pupil is
recorded on the first day of term. Input and store the weights and names recorded for a class of
30 pupils. You must store the weights in a one-dimensional array and the names in another onedimensional array. All the weights must be validated on entry and any invalid weights rejected. You
must decide your own validation rules. You may assume that the pupils’ names are unique. Output
the names and weights of the pupils in the class.
TASK 2
The weight, in kilograms, of each pupil is recorded again on the last day of term. Calculate and store
the difference in weight for each pupil.
TASK 3
For those pupils who have a difference in weight of more than 2.5 kilograms, output, with a suitable
message, the pupil’s name, the difference in weight and whether this is a rise or a fall.
Your program must include appropriate prompts for the entry of data. Error messages and other
outputs need to be set out clearly and understandably. All variables, constants and other identifiers
must have meaningful names. Each task must be fully tested.